Refactor rewind helpers and decouple presentation instrumentation
